* commit 'a959b26e9d8ad25e1db28f9ec6a88d17aed08a25':
  bcmdhd: Fix 64-bit address casting

+++ bcmdhd/wpa_supplicant_8_lib/driver_cmd_nl80211.c
@@ -86,7 +86,7 @@ int wpa_driver_nl80211_driver_cmd(void *priv, char *cmd, char *buf,
 		os_strlcpy(ifr.ifr_name, bss->ifname, IFNAMSIZ);
 
 #ifdef BCMDHD_64_BIT_IPC
-		priv_cmd.bufaddr = (u64)buf;
+		priv_cmd.bufaddr = (u64)(uintptr_t)buf;
 #else
 		priv_cmd.bufaddr = buf;
 #endif
